refactor: separate swapper api methods (#4842)
